initial refactor to showcase
