+++ demos/examples/showcase/ik_servo.rs
@@ -13,16 +13,16 @@
 //! result. The headline is the math cost and its headroom under the 1 ms budget, not a claim of
 //! hard real-time on a desktop OS.
 //!
-//! Streams live to a Rerun viewer; see showcase/viz/README.md for the WSL setup.
-//! Run with: cargo run --release -p multicalc-viz --example ik_servo
+//! Streams live to a Rerun viewer; see demos/README.md for the WSL setup.
+//! Run with: cargo run --release -p multicalc-demos --example ik_servo
 
 #![allow(clippy::unwrap_used, clippy::expect_used, clippy::panic)]
 
 use multicalc::LevenbergMarquardt;
 use multicalc::numerical_derivative::autodiff::AutoDiffMulti;
 use multicalc::scalar::{Numeric, VectorFn};
-use multicalc_viz::loop_util::{LatencyRing, Pacer};
-use multicalc_viz::{RerunSink, Rgba, VizError, VizSink};
+use multicalc_demos::loop_util::{LatencyRing, Pacer};
+use multicalc_demos::{RerunSink, Rgba, VizError, VizSink};
 use std::collections::VecDeque;
 use std::f64::consts::TAU;
 use std::time::Instant;
@@ -126,11 +126,11 @@ fn main() -> Result<(), VizError> {
     if cfg!(debug_assertions) {
         eprintln!(
             "WARNING: debug build — timing numbers are meaningless. \
-             Re-run with: cargo run --release -p multicalc-viz --example ik_servo"
+             Re-run with: cargo run --release -p multicalc-demos --example ik_servo"
         );
     }
 
-    let mut rr = RerunSink::live("multicalc-viz/ik-servo")?;
+    let mut rr = RerunSink::live("multicalc-demos/ik-servo")?;
 
     // Statics: stamp at tick 0 so they forward-fill across the run (see rerun-viz-gotchas).
     rr.set_sequence("tick", 0);
