+++ demos/examples/showcase/support/ferris_outline.rs
