+++ demos/src/csv_sink.rs
