+++ demos/src/loop_util.rs
