+++ demos/src/sink.rs
