--- showcase/viz/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "multicalc-viz"
-version.workspace = true
-edition.workspace = true
-# Rerun requires Rust 1.92, above the workspace 1.85 floor, so this crate sets its own.
-rust-version = "1.92"
-license.workspace = true
-publish = false
-
-[dependencies]
-multicalc = { path = "../../crates/multicalc" }
-# Pinned to match the paired Rerun viewer version. Only the logging SDK is needed: `spawn()`
-# launches the external viewer on PATH, so the in-process `native_viewer` is not compiled in.
-rerun = { version = "=0.33.1", default-features = false, features = ["sdk"] }
-
-[lints]
-workspace = true
